Expose Avro schema via Intake dataset.discover() metadata. - Avro schema if exists can be returned from `dataset.discover()['metadata']` method.
