+++ intake_dal/dal_online.py
@@ -1,10 +1,8 @@
 import base64
 import io
-import json
 import urllib.parse
-from collections import Iterable
 from http import HTTPStatus
-from typing import Dict, Optional, Union
+from typing import Dict
 from urllib.parse import ParseResult, urldefrag, urlparse  # noqa: F401
 
 import numpy as np
@@ -42,25 +40,10 @@ class DalOnlineSource(DataSource):
         self._key_name = parse_result.fragment
         if self._key_name == "":
             raise ValueError(f"key_name expected in URL fragment of {urlpath}")
+
+        self._canonical_name = None  # set in _get_schema()
         self._key_value = key
-        self._canonical_name = None  # _get_schema() sets this
         super().__init__(metadata=metadata)
-
-    def _get_schema(self) -> Schema:
-        if self._canonical_name is None:
-            self._canonical_name = self.metadata["canonical_name"]  # an intake-dal feature
-            # TODO(talebz): Getting avro schema should be promoted to Intake
-            self._avro_schema = _get_avro(self, self._canonical_name)
-            self._schema_dtypes = _avro_to_dtype(self._avro_schema)
-            self._dtypes = {k: str(v) for (k, v) in self._schema_dtypes.items()}
-
-        return Schema(
-            datashape=None,
-            dtype=self._dtypes,
-            shape=(None, len(self._dtypes)),
-            npartitions=1,  # This data is not partitioned, so there is only one partition
-            extra_metadata={"canonical_name": self._canonical_name, "urlpath": self._urlpath},
-        )
 
     def write(self, df: pd.DataFrame):
         self._get_schema()
@@ -78,6 +61,25 @@ class DalOnlineSource(DataSource):
 
     def _close(self):
         pass
+
+    def _get_schema(self) -> Schema:
+        if self._canonical_name is None:
+            self._canonical_name = self.metadata["canonical_name"]
+            self._avro_schema = self.metadata["avro_schema"]
+            self._dtypes = self.metadata["dtypes"]
+            self._storage_mode = self.metadata["storage_mode"]
+
+        return Schema(
+            datashape=None,
+            dtype=self._dtypes,
+            shape=(None, len(self._dtypes)),
+            npartitions=1,  # This data is not partitioned, so there is only one partition
+            extra_metadata={
+                "canonical_name": self._canonical_name,
+                "storage_mode": self._storage_mode,
+                "avro_schema": self._avro_schema,
+            },
+        )
 
 
 AVRO_DATA_SETS_PATH = "avro-data-sets"
@@ -110,84 +112,3 @@ def serialize_panda_df_to_str(df: pd.DataFrame, schema: Dict) -> str:
 
 def deserialize_avro_str_to_pandas(avro_str: str, schema: dict = None) -> pd.DataFrame:
     return pandavro.from_avro(io.BytesIO(base64.b64decode(avro_str)), schema)
-
-
-def _get_avro(source: DataSource, canonical_name: str) -> Optional[Dict]:
-    data_schema_entry = _get_metadata_schema(source)
-
-    if "kafka_schema_registry" in data_schema_entry:
-        # TODO(talebz): check data_schema_entry for kafka_schema_registry.  If exists then query Kafka Schema Registry
-        raise NotImplementedError(
-            "kafka_schema_registry integration not yet supported.  "
-            f"Please put schema as {canonical_name}: > JSON of avro schema"
-        )
-
-    if canonical_name in data_schema_entry:
-        return json.loads(data_schema_entry[canonical_name])
-    else:
-        return None
-
-
-def _get_metadata_schema(source: DataSource) -> Dict:
-    if "data_schema" in source.metadata:
-        return source.metadata["data_schema"]
-    elif source.cat:
-        return _get_metadata_schema(source.cat)
-
-
-# TODO(talebz): ensure this is comprehensive with unit tests!
-def _avro_to_dtype(schema: Dict) -> Dict:
-    field_schemas = {f["name"]: f["type"] for f in schema["fields"]}
-    avro_type_to_dtype = {
-        tuple(sorted(["type", "long", "logicalType", "timestamp-millis"])): np.dtype("datetime64"),
-        tuple(sorted(["type", "long", "logicalType", "timestamp-micros"])): np.dtype("datetime64"),
-        tuple(sorted(["null", "int"])): np.dtype("int32"),
-        tuple(sorted(["null", "long"])): np.dtype("int32"),
-        tuple(sorted(["type", "int", "unsigned", "True"])): np.dtype("uint32"),
-        tuple(sorted(["type", "long", "unsigned", "True"])): np.dtype("int64"),
-        tuple(["long"]): np.dtype("int64"),
-        tuple(["int"]): np.dtype("int32"),
-        tuple(["float"]): np.dtype("float32"),
-        tuple(["double"]): np.dtype("float64"),
-        tuple(["boolean"]): np.dtype("bool"),
-        tuple(["string"]): np.dtype("object"),
-    }
-
-    def to_lookup(avro_type: Union[str, list]) -> tuple:
-        if isinstance(avro_type, str):
-            return tuple([avro_type])
-        elif isinstance(avro_type, list):
-            return tuple(sorted(_flatten(avro_type)))
-
-    ret = {}
-    for (k, v) in field_schemas.items():
-        lookup = to_lookup(v)
-        if lookup in avro_type_to_dtype:
-            ret[k] = avro_type_to_dtype[lookup]
-        elif "null" in lookup:
-            list_lookup = list(lookup)
-            list_lookup.remove("null")
-            new_lookup = tuple(list_lookup)
-            if new_lookup in avro_type_to_dtype:
-                ret[k] = avro_type_to_dtype[new_lookup]
-            else:
-                raise ValueError(f"{lookup} to pandas type not supported in {schema}")
-        else:
-            raise ValueError(f"{lookup} to pandas type not supported in {schema}")
-
-    return ret
-
-
-def _flatten(ls: Iterable) -> Iterable:
-    def iter_ls():
-        if isinstance(ls, dict):
-            return ls.items()
-        else:
-            return ls
-
-    for i in iter_ls():
-        if isinstance(i, Iterable) and not isinstance(i, str):
-            for sub_collection in _flatten(i):
-                yield sub_collection
-        else:
-            yield i
